% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval-save.r
\name{eval.ret}
\alias{eval.ret}
\title{Load a previously saved result}
\usage{
eval.ret(name)
}
\arguments{
\item{name}{Name given to the result.}
}
\description{
Load a previously saved result
}
\examples{
x <- eval.save(f(3),"f")
eval.ret("f")
}

