% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/globals.r
\name{eval.save.dir}
\alias{eval.save.dir}
\title{Directory where evaluations are saved}
\usage{
eval.save.dir(dir)
}
\arguments{
\item{dir}{Directory where evaluation results should be saved.}
}
\value{
The current directory where evaluations are saved.
}
\description{
Directory where evaluations are saved
}

