% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval-save.r
\name{eval.save}
\alias{eval.save}
\title{Evaluate an expression and save the result}
\usage{
eval.save(expr, name, redo = F, verbose = T)
}
\arguments{
\item{expr}{Expression to be evaluated.}

\item{name}{Unique name for later accessing the saved result.}

\item{redo}{If the expression has been previously evaluated and saved
it will normally not be re-evaluated.  This can be overridden
by setting this parameter to TRUE (Default: FALSE).}

\item{verbose}{Print status and error messages (Default: TRUE).}
}
\value{
The result of the expression evaluation.
}
\description{
Evaluate an expression and save the result
}
\examples{
f <- function(v) {
  cat(paste("f(", v, ") = ", x <- mean(runif(v)), "\\n", sep=""));
  x
}

x <- eval.save(f(3),"f")
}

