% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval-save.r
\name{eval.saved}
\alias{eval.saved}
\title{Check if a result has been saved}
\usage{
eval.saved(name)
}
\arguments{
\item{name}{Name given to the result.}
}
\description{
Check if a result has been saved
}
\examples{
x <- eval.save(f(3), "f")
eval.saved("f")
eval.saved("g")
}

